+++ CHANGELOG.rst
@@ -1,8 +1,9 @@
-Version 0.1.1 - Development
+Version 0.1.0 - Development
 =======================================
 
 * First official alpha release.
+* Working for python 2.6 and up, including python 3.
 * Five different type of applications: HTTP server, RPC server, distributed task queue,
   asynchronous test suite and asynchronous shell.
 * 4 examples spanning all applications.
-* **27 regression tests**
+* **35 regression tests**